tech(base): add generator file and add commander
