+++ package.json
@@ -1,18 +1,18 @@
 {
   "name": "@chef/component",
   "version": "1.0.0",
-  "description": "Create Components",
+  "description": "create components skeleton",
   "main": "index.js",
   "scripts": {
     "clean": "rm -rf blocks",
-    "create": "npm run clean && node index.js"
+    "create": "npm run clean && node generator.js"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ramsunvtech/chef-component.git"
   },
   "bin": {
-    "react-chef": "./recipes/index.js"
+    "@chef-component": "./generator.js"
   },
   "keywords": [
     "react-components",
@@ -25,6 +25,7 @@
   },
   "homepage": "https://github.com/ramsunvtech/chef-component#readme",
   "dependencies": {
+    "commander": "^9.0.0",
     "event-stream": "^4.0.1",
     "shelljs": "^0.8.5"
   }
